media: platform: laser_focus: Correct Laser_log_cnt size.

+++ drivers/media/platform/msm/camera_v2/sensor/laser_focus/laser_HEPTAGON/src/HPTG_shipping_func.c
@@ -36,7 +36,7 @@ static uint16_t debug_raw_range = 0;
 static uint16_t debug_raw_confidence = 0;
 
 extern int g_factory;
-extern int Laser_log_cnt;
+extern uint16_t Laser_log_cnt;
 extern int proc_log_cnt;
 extern uint16_t Settings[NUMBER_OF_SETTINGS];
